Move the content of stochtrace_funm.py to funm.py because all internal logic is the same
