+++ matfree/funm.py
@@ -1,5 +1,22 @@
 """Matrix-free implementations of functions of matrices.
 
+This includes matrix-function-vector products
+
+$$
+(f, v, p) \\mapsto f(A(p))v
+$$
+
+as well as matrix-function extensions for stochastic trace estimation,
+which provide
+
+$$
+(f, v, p) \\mapsto v^\\top f(A(p))v.
+$$
+
+Plug these integrands into
+[matfree.stochtrace.estimator][matfree.stochtrace.estimator].
+
+
 Examples
 --------
 >>> import jax.random
@@ -20,7 +37,8 @@ Examples
 Array([-4.1, -1.3, -2.2, -2.1, -1.2, -3.3, -0.2,  0.3,  0.7,  0.9],      dtype=float32)
 """
 
-from matfree.backend import containers, control_flow, func, linalg, np
+from matfree import decomp
+from matfree.backend import containers, control_flow, func, linalg, np, tree_util
 from matfree.backend.typing import Array, Callable
 
 
@@ -128,8 +146,118 @@ def funm_lanczos_sym(dense_funm: Callable, tridiag_sym: Callable, /) -> Callable
     return estimate
 
 
+def integrand_funm_sym_logdet(order, matvec, /):
+    """Construct the integrand for the log-determinant.
+
+    This function assumes a symmetric, positive definite matrix.
+    """
+    return integrand_funm_sym(np.log, order, matvec)
+
+
+def integrand_funm_sym(matfun, order, matvec, /):
+    """Construct the integrand for matrix-function-trace estimation.
+
+    This function assumes a symmetric matrix.
+    """
+    # todo: if we ask the user to flatten their matvecs,
+    #  then we can give this code the same API as funm_lanczos_sym.
+    dense_funm = dense_funm_sym_eigh(matfun)
+
+    def quadform(v0, *parameters):
+        v0_flat, v_unflatten = tree_util.ravel_pytree(v0)
+        length = linalg.vector_norm(v0_flat)
+        v0_flat /= length
+
+        def matvec_flat(v_flat, *p):
+            v = v_unflatten(v_flat)
+            Av = matvec(v, *p)
+            flat, unflatten = tree_util.ravel_pytree(Av)
+            return flat
+
+        algorithm = decomp.tridiag_sym(matvec_flat, order)
+        (_, (diag, off_diag)), _ = algorithm(v0_flat, *parameters)
+
+        dense = _todense_tridiag_sym(diag, off_diag)
+        fA = dense_funm(dense)
+        e1 = np.eye(len(fA))[0, :]
+        return length**2 * linalg.inner(e1, fA @ e1)
+
+    return quadform
+
+
+def integrand_funm_product_logdet(depth, matvec, vecmat, /):
+    r"""Construct the integrand for the log-determinant of a matrix-product.
+
+    Here, "product" refers to $X = A^\top A$.
+    """
+    return integrand_funm_product(np.log, depth, matvec, vecmat)
+
+
+def integrand_funm_product_schatten_norm(power, depth, matvec, vecmat, /):
+    r"""Construct the integrand for the $p$-th power of the Schatten-p norm."""
+
+    def matfun(x):
+        """Matrix-function for Schatten-p norms."""
+        return x ** (power / 2)
+
+    return integrand_funm_product(matfun, depth, matvec, vecmat)
+
+
+def integrand_funm_product(matfun, depth, matvec, vecmat, /):
+    r"""Construct the integrand for matrix-function-trace estimation.
+
+    Instead of the trace of a function of a matrix,
+    compute the trace of a function of the product of matrices.
+    Here, "product" refers to $X = A^\top A$.
+    """
+
+    def quadform(v0, *parameters):
+        v0_flat, v_unflatten = tree_util.ravel_pytree(v0)
+        length = linalg.vector_norm(v0_flat)
+        v0_flat /= length
+
+        def matvec_flat(v_flat, *p):
+            v = v_unflatten(v_flat)
+            Av = matvec(v, *p)
+            flat, unflatten = tree_util.ravel_pytree(Av)
+            return flat, tree_util.partial_pytree(unflatten)
+
+        w0_flat, w_unflatten = func.eval_shape(matvec_flat, v0_flat)
+        matrix_shape = (*np.shape(w0_flat), *np.shape(v0_flat))
+
+        def vecmat_flat(w_flat):
+            w = w_unflatten(w_flat)
+            wA = vecmat(w, *parameters)
+            return tree_util.ravel_pytree(wA)[0]
+
+        # Decompose into orthogonal-bidiag-orthogonal
+        algorithm = decomp.bidiag(
+            lambda v: matvec_flat(v)[0], vecmat_flat, depth, matrix_shape=matrix_shape
+        )
+        output = algorithm(v0_flat, *parameters)
+        u, (d, e), vt, _ = output
+
+        # Compute SVD of factorisation
+        B = _todense_bidiag(d, e)
+
+        # todo: turn the following lines into dense_funm_svd()
+        _, S, Vt = linalg.svd(B, full_matrices=False)
+
+        # Since Q orthogonal (orthonormal) to v0, Q v = Q[0],
+        # and therefore (Q v)^T f(D) (Qv) = Q[0] * f(diag) * Q[0]
+        eigvals, eigvecs = S**2, Vt.T
+        fx_eigvals = func.vmap(matfun)(eigvals)
+        return length**2 * linalg.inner(eigvecs[0, :], fx_eigvals * eigvecs[0, :])
+
+    return quadform
+
+
 def dense_funm_sym_eigh(matfun):
-    """Implement dense matrix-functions via symmetric eigendecompositions."""
+    """Implement dense matrix-functions via symmetric eigendecompositions.
+
+    Use it to construct one of the matrix-free matrix-function implementations,
+    e.g. [matfree.funm.funm_lanczos_sym][matfree.funm.funm_lanczos_sym].
+    """
 
     def fun(dense_matrix):
         eigvals, eigvecs = linalg.eigh(dense_matrix)
@@ -140,7 +268,11 @@ def dense_funm_sym_eigh(matfun):
 
 
 def dense_funm_schur(matfun):
-    """Implement dense matrix-functions via symmetric Schur decompositions."""
+    """Implement dense matrix-functions via symmetric Schur decompositions.
+
+    Use it to construct one of the matrix-free matrix-function implementations,
+    e.g. [matfree.funm.funm_lanczos_sym][matfree.funm.funm_lanczos_sym].
+    """
 
     def fun(dense_matrix):
         return linalg.funm_schur(dense_matrix, matfun)
@@ -148,11 +280,19 @@ def dense_funm_schur(matfun):
     return fun
 
 
+# todo: if we move this logic to the decomposition algorithms
+#  (e.g. with a materalize=True flag in the decomposition construction),
+#  then all trace_of_funm implementation reduce to very few lines.
+
+
 def _todense_tridiag_sym(diag, off_diag):
-    # todo: once jax supports eigh_tridiagonal(eigvals_only=False),
-    #  use it here. Until then: an eigen-decomposition of size (order + 1)
-    #  does not hurt too much...
     diag = linalg.diagonal_matrix(diag)
     offdiag1 = linalg.diagonal_matrix(off_diag, -1)
     offdiag2 = linalg.diagonal_matrix(off_diag, 1)
     return diag + offdiag1 + offdiag2
+
+
+def _todense_bidiag(d, e):
+    diag = linalg.diagonal_matrix(d)
+    offdiag = linalg.diagonal_matrix(e, 1)
+    return diag + offdiag
