+++ tests/test_funm/test_integrand_funm_schatten_norm.py
@@ -1,6 +1,6 @@
 """Test stochastic Lanczos quadrature for Schatten-p-norms."""
 
-from matfree import stochtrace, stochtrace_funm, test_util
+from matfree import funm, stochtrace, test_util
 from matfree.backend import linalg, np, prng, testing
 
 
@@ -27,7 +27,7 @@ def test_schatten_norm(A, order, power):
     _, ncols = np.shape(A)
     args_like = np.ones((ncols,), dtype=float)
     sampler = stochtrace.sampler_normal(args_like, num=500)
-    integrand = stochtrace_funm.integrand_funm_product_schatten_norm(
+    integrand = funm.integrand_funm_product_schatten_norm(
         power, order, lambda v: A @ v, lambda v: A.T @ v
     )
     estimate = stochtrace.estimator(integrand, sampler)
